fix: spleat components on parts
