--- src/.eslintrc.json
@@ -1,34 +1,0 @@
-{
-  "plugins": [
-    "prettier", 
-    "import", 
-    "@typescript-eslint"
-  ],
-  "env": {
-    "es6": true,
-    "browser": true,
-    "node": true
-  },
-  "extends": [
-    "plugin:@typescript-eslint/recommended",
-    "eslint:recommended",
-    "plugin:prettier/recommended",
-    "prettier"
-  ],
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-      "ecmaVersion": "2020",
-      "sourceType": "module",
-      "project": ["tsconfig.json"]
-  },
-  "rules": {
-      "no-debugger": "off",
-      "no-console": 0,
-      "class-methods-use-this": "off",
-      "no-unused-vars": "off",
-      "@typescript-eslint/no-inferrable-types": 1,
-      "@typescript-eslint/no-var-requires": 0,
-      "@typescript-eslint/no-explicit-any": 2,
-      "prettier/prettier": ["error", { "endOfLine": "auto" }]
-  }
-}
