--- src/.prettierrc
@@ -1,10 +1,0 @@
-{
-  "trailingComma": "es5",
-  "tabWidth": 2,
-  "semi": true,
-  "singleQuote": true,
-  "printWidth": 120,
-  "bracketSpacing": true,
-  "bracketSameLine": false,
-  "arrowParens": "always"
-}
