+++ src/Components/Main page/Nav.ts
