--- src/README.md
@@ -1,1 +1,0 @@
-# rs-clone
