+++ src/assets/icons/coins.png
