+++ src/index.html
@@ -8,122 +8,7 @@
   <title>Rs Clone</title>
 </head>
 
-<body class="font-sans h-screen min-w-min overflow-x-hidden">
-<div class="container mx-auto">
-  <header class="header fixed bottom-0 left-0 right-0 z-40 bg-white">
-    <nav class="nav relative shadow-top">
-      <div class="nav__menu px-5 py-3">
-        <ul class="nav__list flex flex-row justify-between text-xs">
-          <li class="nav__item">
-            <a href="#" class="nav__link flex flex-col items-center">
-              <img class="nav__icon w-6 h-6" src="../src/assets/icons/home.png" alt="home-icon">
-              <span class="nav__name">Главная</span>
-            </a>
-          </li>
-          <li class="nav__item">
-            <a href="#" class="nav__link flex flex-col items-center">
-              <img class="nav__icon w-6 h-6" src="../src/assets/icons/event.png" alt="event-icon">
-              <span class="nav__name">События</span>
-            </a>
-          </li>
-          <li class="nav__item">
-            <a href="#" class="nav__link flex flex-col items-center">
-              <img class="nav__icon w-6 h-6" src="../src/assets/icons/calendar.png" alt="calendar-icon">
-              <span class="nav__name">Планы</span>
-            </a>
-          </li>
-          <li class="nav__item">
-            <a href="#" class="nav__link flex flex-col items-center">
-              <img class="nav__icon w-6 h-6" src="../src/assets/icons/pie-chart.png" alt="pie-icon">
-              <span class="nav__name">Статистика</span>
-            </a>
-          </li>
-        </ul>
-      </div>
-      <button class="nav__button absolute bottom-14 right-0 w-10 h-10 p-2 rounded-full bg-blue">
-      <img src="./assets/icons/plus.svg" alt="plus">
-    </button>
-    </nav>
-  </header>
-
-    <main class="main">        
-        <section class="info">
-          <h1 class="info__title py-3 font-bold text-lg">Мой автомобиль</h1>
-          <div class="info__car car bg-slate rounded-lg shadow-md mb-2 p-4 grid grid-areas-layout gap-x-2">
-            <h3 class="car__name font-bold text-sm text-right grid-in-name">Chevrolet Aveo</h3>
-            <p class="car__year text-xs text-right grid-in-year">2007 г.в.</з>
-              <div class="car__mileage grid-in-mileage">
-                <div class="car__mileage-total leading-3 mb-2">
-                  <h3 class="car__mileage-total_amount font-bold text-sm">
-                    127 000<span> км</span>
-                  </h3>
-                  <span class="text-xs">Пробег</span>
-                </div>
-                <div class="car__mileage-average leading-3">
-                  <h3 class="car__mileage-average_amount font-bold text-sm">
-                    8 500<span> км</span>
-                  </h3>
-                  <span class="text-xs">Среднегодовой пробег</span>
-                </div>
-              </div>
-              <div class="car__year-image">
-                <img src="./assets/images/chevrolet.png" alt="" class="car__image">
-              </div>
-          </div>
-        </section>
-
-        <section class="plans">
-          <h2 class="plans__title font-medium text-sm mb-2">Ближайшие планы</h2>
-          <ul class="plans__list grid gap-y-3">
-            <li class="plans__item relative bg-slate rounded-lg flex justify-between gap-x-2 items-center py-2 pl-4 shadow-md">
-              <div class="plans__bar absolute bottom-14 right-0 bg-blue text-white text-xs px-3 py-1 rounded-md">Через 10 дней</div>
-              <div class="plans__item-and-icon flex items-center mr-2">
-                <img src="./assets/icons/car-oil.png" class="plans__image w-7 h-7 mr-4" alt="oil-icon">
-                <div class="plans__text">
-                  <h3 class="plans__title text-sm font-bold mb-1 leading-3">Замена масла</h3>
-                  <span class="plans__addition text-xs leading-3 inline-block">Масло моторное SHELL HELIX HX8 SW-40 4л.
-                  </span>
-                </div>
-              </div>
-              <div class="plans__date text-xs font-bold">25 дек. 2022 г.</div>
-            </li>
-            <li class="plans__item bg-slate rounded-lg flex justify-between gap-x-2 items-center py-2 pl-4 pr-2 shadow-md">
-              <div class="plans__item-and-icon flex">
-                <img src="./assets/icons/car-repair.png" class="plans__image w-7 h-7 mr-4" alt="oil-icon">
-                <h3 class="plans__title text-sm font-bold mb-1 leading-3">Диагностика подвески</h3>
-              </div>
-              <button class="plans__button button text-xs bg-blue text-white px-3 py-1 rounded-md">Добавить</button>
-            </li>
-            <li class="plans__item bg-slate rounded-lg flex justify-between gap-x-2 items-center py-2 pl-4 pr-2 shadow-md">
-              <div class="plans__item-and-icon flex">
-                <img src="./assets/icons/gear.png" class="plans__image w-7 h-7 mr-4" alt="oil-icon">
-                <h3 class="plans__title text-sm font-bold mb-1 leading-3">Замена фильтров</h3>
-              </div>
-              <button class="plans__button button text-xs bg-blue text-white px-3 py-1 rounded-md">Добавить</button>
-            </li>
-          </ul>
-        </section>
-
-        <section class="events py-2 px-4 rounded-lg shadow-top absolute bottom-16 bg-white w-full left-0 min-w-min">
-          <h2 class="events__title font-medium text-xxs">Последние события</h2>
-          <ul class="events__list grid gap-y-3">
-            <li class="events__item flex justify-between items-center py-2 pl-2">
-              <div class="events__item-and-icon flex items-center mr-2">
-                <img src="./assets/icons/funnel.png" class="plans__image w-7 h-7 mr-4" alt="funnell-icon">
-                <div class="events__text flex flex-col">
-                  <h3 class="events__title text-sm font-bold mb-1 leading-4">Замена топливного фильтра</h3>
-                  <span class="events__addition text-xxs leading-3 inline-block">Вместе с фильтром. Автосервис "Перекресток"
-                  </span>
-                </div>
-              </div>
-              <div class="events__date-and-price self-end">
-                <div class="events__price text-xs font-bold mb-1">3 900 ₽</div>
-                <div class="events__date text-xxs leading-3">01 нояб. 2022 г.</div>
-              </div>
-            </li>
-          </ul>
-        </section>
-    </main>
-  </div>  
+<body class="font-sans h-screen min-w-min overflow-x-hidden inset-0">
+  <div id="root"></div>
 </body>
 </html>
