+++ src/index.ts
@@ -1,4 +1,11 @@
 import './index.html';
 import './main.css';
+import { App } from './App/App';
 
-console.log('TS работает. CSS тоже.');
+const root = document.querySelector('#root');
+const app = new App();
+
+if (!root) throw new Error('Failed to load container element');
+else {
+  root.innerHTML = app.render();
+}
