--- src/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "rs-clone",
-  "version": "0.0.1",
-  "description": "RS School final task.",
-  "main": "index.js",
-  "scripts": {
-    "start": "webpack serve",
-    "build": "webpack --node-env=production",
-    "lint": "eslint . --ext .ts --fix"
-  },
-  "keywords": [],
-  "author": "",
-  "devDependencies": {
-    "@savvywombat/tailwindcss-grid-areas": "^3.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.46.1",
-    "@typescript-eslint/parser": "^5.48.2",
-    "autoprefixer": "^10.4.13",
-    "css-loader": "^6.7.3",
-    "eslint": "^8.29.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-webpack-plugin": "^3.2.0",
-    "html-loader": "^4.2.0",
-    "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.7.2",
-    "postcss": "^8.4.21",
-    "prettier": "2.8.1",
-    "sass": "^1.56.2",
-    "sass-loader": "^13.2.0",
-    "style-loader": "^3.3.1",
-    "tailwindcss": "^3.2.4",
-    "ts-loader": "^9.4.2",
-    "typescript": "^4.9.4",
-    "webpack": "^5.75.0",
-    "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.11.1"
-  },
-  "dependencies": {
-    "postcss-loader": "^7.0.2"
-  }
-}
