--- src/src/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Rs Clone</title>
-</head>
-
-<body class="font-sans h-screen min-w-min overflow-x-hidden">
-  <div id="root"></div>
-</body>
-</html>
