--- src/webpack.config.js
@@ -1,70 +1,0 @@
-const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const EslingPlugin = require('eslint-webpack-plugin');
-
-const mode = process.env.NODE_ENV || 'development';
-const devMode = mode === 'development';
-const devtool = devMode ? 'source-map' : undefined;
-
-module.exports = {
-  mode,
-  devtool,
-  devServer: {
-    port: 8080,
-    open: true,
-  },
-  entry: path.resolve(__dirname, 'src', 'index'),
-  resolve: {
-    extensions: ['.ts', '.js'],
-  },
-  output: {
-    path: path.resolve(__dirname, 'dist'),
-    clean: true,
-    filename: 'index.[contenthash].js',
-    assetModuleFilename: 'assets/[name][ext]',
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: path.resolve(__dirname, 'src', 'index.html'),
-      inject: 'body',
-    }),
-    new MiniCssExtractPlugin({
-      filename: 'index.[contenthash].css',
-    }),
-    new EslingPlugin({
-      extensions: 'ts',
-    }),
-  ],
-  module: {
-    rules: [
-      {
-        test: /.ts?$/i,
-        use: ['ts-loader'],
-        exclude: /node_modules/,
-      },
-      {
-        test: /\.html$/i,
-        loader: 'html-loader',
-      },
-      {
-        test: /\.(c|sa|sc)ss$/i,
-        use: [devMode ? 'style-loader' : MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader'],
-      },
-      {
-        test: /\.woff2?$/i,
-        type: 'asset/resource',
-        generator: {
-          filename: 'fonts/[name][ext]',
-        },
-      },
-      {
-        test: /\.(jpe?g|png|webp|gif|svg)$/i,
-        type: 'asset/resource',
-      },
-    ],
-  },
-  experiments: {
-    topLevelAwait: true,
-  },
-};
