+++ webpack.config.js
@@ -21,7 +21,7 @@ module.exports = {
   output: {
     path: path.resolve(__dirname, 'dist'),
     clean: true,
-    filename: 'index.[contenthash].js',
+    filename: 'index.js',
     assetModuleFilename: 'assets/[name][ext]',
   },
   plugins: [
